+++ mem-filter/const.go
@@ -1,0 +1,7 @@
+package filter
+
+// [mem_filter]
+const (
+	keyMaxCapacity     = "mem_filter.max_capacity"
+	defaultMaxCapacity = 32
+)