+++ mem-reader/const.go
@@ -1,0 +1,7 @@
+package reader
+
+// [mem_reader]
+const (
+	keyMaxCapacity     = "mem_reader.max_capacity"
+	defaultMaxCapacity = 32
+)